services/mock-notification-receiver/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via starlette
click==8.3.1
    # via uvicorn
fastapi==0.128.0
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   pydantic-settings
sniffio==1.3.1
    # via anyio
starlette==0.50.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.40.0
    # via -r requirements.in
